Deploy version 0.7.0

+++ Assets/Plugins/iOS/Wikitude/WikitudeMiraSDK.framework/Headers/WTPoint3D.h
@@ -7,9 +7,9 @@
 //
 
 struct WTPoint3D{
-    CGFloat x;
-    CGFloat y;
-    CGFloat z;
+    float x;
+    float y;
+    float z;
 };
 
 typedef struct WTPoint3D WTPoint3D;
